[#45] 불필요한 이메일 로그인 관련 로직 제거
